chore: code review tweaks
