+++ v3/src/components/map/map-registration.ts
@@ -9,7 +9,8 @@ import {
 import { registerTileComponentInfo } from "../../models/tiles/tile-component-info"
 import { ITileLikeModel, registerTileContentInfo } from "../../models/tiles/tile-content-info"
 import { t } from "../../utilities/translation/translate"
-import {registerV2TileImporter} from "../../v2/codap-v2-tile-importers"
+import { registerV2TileExporter } from "../../v2/codap-v2-tile-exporters"
+import { registerV2TileImporter } from "../../v2/codap-v2-tile-importers"
 import { ComponentTitleBar } from "../component-title-bar"
 import {
   AttributeDescriptionsMapSnapshot, kDataConfigurationType
@@ -19,18 +20,17 @@ import { MapInspector } from "./components/map-inspector"
 import {kMapIdPrefix, kMapTileClass, kMapTileType, kV2MapType} from "./map-defs"
 import {kDefaultMapHeight, kDefaultMapWidth, kMapPointLayerType, kMapPolygonLayerType} from "./map-types"
 import { IMapBaseLayerModelSnapshot } from "./models/map-base-layer-model"
-import { IMapPointLayerModelSnapshot } from "./models/map-point-layer-model"
-import { IMapPolygonLayerModelSnapshot } from "./models/map-polygon-layer-model"
 import {
   createMapContentModel, IMapModelContentSnapshot, isMapContentModel, MapContentModel
 } from "./models/map-content-model"
+import { IMapPointLayerModelSnapshot } from "./models/map-point-layer-model"
+import { IMapPolygonLayerModelSnapshot } from "./models/map-polygon-layer-model"
 import { MapFilterFormulaAdapter } from "./models/map-filter-formula-adapter"
 import {
   boundaryAttributeFromDataSet, datasetHasBoundaryData, datasetHasLatLongData, latLongAttributesFromDataSet
 } from "./utilities/map-utils"
-import {v2MapImporter} from "./v2-map-importer"
-import { v2MapExporter } from "./map-V2-exporter"
-import { registerV2TileExporter } from "../../v2/codap-v2-tile-exporters"
+import { v2MapExporter } from "./v2-map-exporter"
+import { v2MapImporter } from "./v2-map-importer"
 
 MapFilterFormulaAdapter.register()
