+++ v3/src/components/map/v2-map-importer.test.ts
@@ -4,12 +4,15 @@ import { SharedModelDocumentManager } from "../../models/document/shared-model-d
 import { ITileModelSnapshotIn } from "../../models/tiles/tile-model"
 import { safeJsonParse } from "../../utilities/js-utils"
 import { CodapV2Document } from "../../v2/codap-v2-document"
-import { ICodapV2DocumentJson, ICodapV2MapLayerStorage, ICodapV2MapPointLayerStorage, ICodapV2MapPolygonLayerStorage,
-          ICodapV2MapStorage, isV2MapCurrentStorage, isV2MapPointLayerStorage, isV2MapPolygonLayerStorage }
-    from "../../v2/codap-v2-types"
+import {
+  ICodapV2DocumentJson, ICodapV2MapCurrentStorage, ICodapV2MapLayerStorage,
+  ICodapV2MapPointLayerStorage, ICodapV2MapPolygonLayerStorage, ICodapV2MapStorage,
+  isV2MapCurrentStorage, isV2MapPointLayerStorage, isV2MapPolygonLayerStorage
+} from "../../v2/codap-v2-types"
 import { isMapContentModel } from "./models/map-content-model"
+import { v2MapExporter } from "./v2-map-exporter"
 import { v2MapImporter } from "./v2-map-importer"
-import { v2MapExporter } from "./map-V2-exporter"
+
 import "./map-registration"
 
 const fs = require("fs")
@@ -120,23 +123,12 @@ describe("imports/exports to current v2 map documents", () => {
     expect(mapModel?.layers.length).toBe(2)
   })
 
-  it("exports a V2 map component and re-imports it correctly", () => {
+  it("imports a V2 map component and re-exports it correctly", () => {
     const importedTile = v2MapImporter({
       v2Component: firstMapComponent(v2Document),
       v2Document,
       insertTile: mockInsertTile
     })
-
-    type IMapModelStorage = {
-      center: { lat: number, lng: number } | [lat: number, lng: number]
-      zoom: number
-      baseMapLayerName: string
-      // TODO_V2_IMPORT: gridMultiplier is not imported at this level
-      // It appears 8,612 times in cfm-shared either here or
-      // inside of the grid object
-      gridMultiplier: number
-      layerModels: ICodapV2MapLayerStorage[]
-    }
 
     expect(importedTile).toBeDefined()
     expect(isMapContentModel(importedTile?.content)).toBe(true)
@@ -147,7 +139,7 @@ describe("imports/exports to current v2 map documents", () => {
     expect(exportedV2Map).toBeDefined()
     expect(exportedV2Map?.type).toBe("DG.MapView")
     const mapStorage = exportedV2Map?.componentStorage as ICodapV2MapStorage
-    let mapModelStorage: IMapModelStorage | undefined
+    let mapModelStorage: Maybe<ICodapV2MapCurrentStorage["mapModelStorage"]>
     let layerModels: ICodapV2MapLayerStorage[] = []
     if (isV2MapCurrentStorage(mapStorage)) {
       mapModelStorage = mapStorage.mapModelStorage
@@ -155,7 +147,6 @@ describe("imports/exports to current v2 map documents", () => {
     }
     expect(mapModelStorage).toBeDefined()
     expect(layerModels.length).toBe(2)
-
 
     const polygonLayer = layerModels[0] as ICodapV2MapPolygonLayerStorage
     const isFirstPolygon = isV2MapPolygonLayerStorage(polygonLayer)
